Move OpenDosh logic to effects
